refactor(graphql): refactor getMetadata call
Closes SWAP-2731
