+++ src/graphql/questionary/fragment.fileMetadata.graphql
@@ -1,0 +1,7 @@
+fragment fileMetadata on FileMetadata {
+  fileId
+  originalFileName
+  mimeType
+  sizeInBytes
+  createdDate
+}
